// client/app/app.ts
export interface simpleExplanation {
    actionValue: string;
}

export interface Button1 {
    buttonStatus: boolean;
}

export interface Button2 {
    buttonStatus: boolean;
}

export interface Button3 {
    buttonStatus: boolean;
}

export interface Form1 {
    inputvalue: string;
    errorMessage: string;
    savedValue: string;
}